Added Navbar components
Introduces 2 new navbar components, Navbar and NavbarWithOpacity.

Navbar is the default and renders with a white background and dark elements

NavbarWithOpacity uses render opaque with light ements and uses hooks to determine the scroll position of the page. When a user scrolls past the offset, the navbar updates to be the default Navbar