Story Setup and Structure
Set up the environment by which login status and tokenization will be handled. Everything is set up with hooks using Context so that we can handle login as state and update UI elements

I also reformatted the way that the Providers are handled in the directory. This will make it easier to develop the connection between the Client and Auth